requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.0

# Data processing
pyspark==3.4.1
apache-airflow==2.7.1
great-expectations==0.17.19
kafka-python==2.0.2

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.20

# ML and NLP
nltk==3.8.1
spacy==3.6.1
transformers==4.33.2
mlflow==2.7.1

# Visualization and Dashboard
streamlit==1.27.0
plotly==5.17.0
wordcloud==0.9.2
matplotlib==3.7.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.1.0
isort==5.12.0
mypy==1.5.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0 